sdk/rust: rename instructions as plural
